Registration updated to allow a new user to be registered and you can now login with the new users data.
The registration validates all the data you entered to make sure it
is not empty, that the username is unique, that the display name is
unique, and that the passwords match.

It them gives you a confirmation and closes the window.

Also added to methods to PlayerList to draw an array of all usernames
and display names.